added fetching types
