+++ src/hooks/useTrending.ts
@@ -1,19 +1,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import APIClient, { FetchResponse} from '../Services/api-client';
+import { Fetching } from '../Pages/Trending';
 
-export interface Fetching{
-    name: string;
-    title: string;
-    poster_path: string;
-    backdrop_path: string;
-    first_air_date: string;
-    release_date: string;
-    media_type: string;
-    id: number;
-    overview: string;
-    vote_average: number;
-    results: []
-  }  
 
 const apiClient = new APIClient<Fetching>('/trending/all/day');
