+++ src/hooks/useUpcoming.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import APIClient, { FetchResponse } from '../Services/api-client'
-import { Fetching } from './useTrending'
+import { Fetching } from '../Pages/Trending'
 
 const apiClient = new APIClient<Fetching>('/movie/upcoming')
