+++ src/types/Fetching.ts
@@ -1,0 +1,13 @@
+export interface Fetching{
+    name: string;
+    title: string;
+    poster_path: string;
+    backdrop_path: string;
+    first_air_date: string;
+    release_date: string;
+    media_type: string;
+    id: number;
+    overview: string;
+    vote_average: number;
+    results: []
+  }
